+++ backend/cozytrain/src/main/java/com/ssafy/cozytrain/api/service/CheckListService.java
@@ -3,5 +3,5 @@ package com.ssafy.cozytrain.api.service;
 import com.ssafy.cozytrain.api.dto.CheckListDto.*;
 
 public interface CheckListService {
-    Long checkListSave(CheckListSaveReq checkListSaveReq);
+    Integer checkListSave(CheckListSaveReq checkListSaveReq);
 }
